@@ -118,14 +118,12 @@ public class VelocityArgumentCommandNode<S, T> extends ArgumentCommandNode<S, St
     if (this == o) {
       return true;
     }
-    if (!(o instanceof VelocityArgumentCommandNode)) {
+    if (!(o instanceof VelocityArgumentCommandNode that)) {
       return false;
     }
-    if (!super.equals(o)) {
+    if (!super.equals(that)) {
       return false;
     }
-
-    final VelocityArgumentCommandNode<?, ?> that = (VelocityArgumentCommandNode<?, ?>) o;
     return this.type.equals(that.type);
   }
 
