@@ -59,14 +59,9 @@ public class SystemChatPacket implements MinecraftPacket {
     component.write(buf);
     if (version.noLessThan(ProtocolVersion.MINECRAFT_1_19_1)) {
       switch (type) {
-        case SYSTEM:
-          buf.writeBoolean(false);
-          break;
-        case GAME_INFO:
-          buf.writeBoolean(true);
-          break;
-        default:
-          throw new IllegalArgumentException("Invalid chat type");
+        case SYSTEM -> buf.writeBoolean(false);
+        case GAME_INFO -> buf.writeBoolean(true);
+        default -> throw new IllegalArgumentException("Invalid chat type");
       }
     } else {
       ProtocolUtils.writeVarInt(buf, type.getId());
